Move SourceI implementations to the forklift.source.sources package
