+++ core/src/main/java/forklift/consumer/Consumer.java
@@ -17,11 +17,11 @@ import forklift.decorators.Response;
 import forklift.message.Header;
 import forklift.producers.ForkliftProducerI;
 import forklift.properties.PropertiesManager;
-import forklift.source.GroupedTopicSource;
-import forklift.source.QueueSource;
 import forklift.source.SourceI;
 import forklift.source.SourceUtil;
-import forklift.source.TopicSource;
+import forklift.source.sources.GroupedTopicSource;
+import forklift.source.sources.QueueSource;
+import forklift.source.sources.TopicSource;
 import forklift.source.decorators.GroupedTopic;
 import forklift.source.decorators.Queue;
 import forklift.source.decorators.Topic;
