+++ core/src/main/java/forklift/source/decorators/Queue.java
@@ -1,7 +1,7 @@
 package forklift.source.decorators;
 
-import forklift.source.QueueSource;
 import forklift.source.SourceType;
+import forklift.source.sources.QueueSource;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
