+++ core/src/main/java/forklift/source/sources/GroupedTopicSource.java
@@ -1,5 +1,6 @@
-package forklift.source;
+package forklift.source.sources;
 
+import forklift.source.SourceI;
 import forklift.source.decorators.GroupedTopic;
 
 import java.util.Objects;
