+++ plugins/replay/src/main/java/forklift/replay/ReplayLogger.java
@@ -5,8 +5,8 @@ import forklift.consumer.Consumer;
 import forklift.consumer.MessageRunnable;
 import forklift.consumer.ProcessStep;
 import forklift.decorators.LifeCycle;
-import forklift.source.QueueSource;
-import forklift.source.TopicSource;
+import forklift.source.sources.QueueSource;
+import forklift.source.sources.TopicSource;
 
 import java.io.File;
 import java.io.FileNotFoundException;
